Taller2/requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
ImageIO==2.37.2
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.4
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
